# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Zero-knowledge, self-hosted password vault API",
        "title": "vaultctl API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/admin/backup": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Backup requires pg_dump (shell access) and is not suitable for HTTP API. Use the CLI instead.",
                "tags": [
                    "Admin"
                ],
                "summary": "Trigger backup",
                "responses": {
                    "501": {
                        "description": "Not implemented — use CLI",
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.ErrorBody"
                        }
                    }
                }
            }
        },
        "/admin/backups": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "List available backup files with size and creation time. Admin only.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "List backups",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.ListBackupsResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.ErrorBody"
                        }
                    },
                    "403": {
                        "description": "Admin required",
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.ErrorBody"
                        }
                    }
                }
            }
        },
        "/auth/invites/redeem": {
            "post": {
                "description": "New user redeems an invite token during registration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invites"
                ],
                "summary": "Redeem invite",
                "parameters": [
                    {
                        "description": "Raw invite token",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.RedeemInviteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.RedeemInviteResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.ErrorBody"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.ErrorBody"
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "description": "Authenticate with email and auth hash, returns JWT tokens and encrypted crypto material",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.LoginResponse"
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.ErrorBody"
                        }
                    },
                    "423": {
                        "description": "Account locked",
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.ErrorBody"
                        }
                    },
                    "429": {
                        "description": "Rate limited",
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.ErrorBody"
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "description": "Revoke the refresh token to end the session",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Logout",
                "parameters": [
                    {
                        "description": "Refresh token to revoke",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.LogoutRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.ErrorBody"
                        }
                    }
                }
            }
        },
        "/auth/password/change": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Change master password and re-encrypt private keys. Requires step-up.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Change master password",
                "parameters": [
                    {
                        "description": "Old and new auth hashes with re-encrypted keys",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.PasswordChangeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.PasswordChangeResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.ErrorBody"
                        }
                    },
                    "403": {
                        "description": "Step-up required",
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.ErrorBody"
                        }
                    }
                }
            }
        },
        "/auth/password/hint": {
            "get": {
                "description": "Returns the password hint for the given email. Returns empty hint for unknown emails (enumeration-safe).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Get password hint",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User email address",
                        "name": "email",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.PasswordHintResponse"
                        }
                    }
                }
            }
        },
        "/auth/prelogin": {
            "get": {
                "description": "Returns salt and KDF parameters needed to derive the auth hash client-side",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Get KDF parameters",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User email address",
                        "name": "email",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.PreloginResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.ErrorBody"
                        }
                    }
                }
            }
        },
        "/auth/recovery/reset": {
            "post": {
                "description": "Reset password after client-side recovery key verification. Revokes all sessions and returns fresh tokens.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Reset password via recovery",
                "parameters": [
                    {
                        "description": "New auth credentials with re-encrypted keys",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.RecoveryResetRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.RecoveryResetResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.ErrorBody"
                        }
                    },
                    "401": {
                        "description": "Unknown email",
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.ErrorBody"
                        }
                    },
                    "429": {
                        "description": "Rate limited",
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.ErrorBody"
                        }
                    }
                }
            }
        },
        "/auth/recovery/verify": {
            "post": {
                "description": "Returns encrypted key material so the client can attempt decryption with its recovery key.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Get recovery material",
                "parameters": [
                    {
                        "description": "Email to recover",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.RecoveryVerifyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.RecoveryVerifyResponse"
                        }
                    },
                    "401": {
                        "description": "Unknown email",
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.ErrorBody"
                        }
                    },
                    "429": {
                        "description": "Rate limited",
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.ErrorBody"
                        }
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "description": "Exchange a valid refresh token for a new access/refresh token pair",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Refresh tokens",
                "parameters": [
                    {
                        "description": "Refresh token",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.RefreshRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.RefreshResponse"
                        }
                    },
                    "401": {
                        "description": "Session expired",
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.ErrorBody"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Create a new user account with client-side derived crypto material",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Register new user",
                "parameters": [
                    {
                        "description": "Registration payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.RegisterResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.ErrorBody"
                        }
                    },
                    "409": {
                        "description": "Email already taken",
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.ErrorBody"
                        }
                    },
                    "429": {
                        "description": "Rate limited",
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.ErrorBody"
                        }
                    }
                }
            }
        },
        "/auth/step-up": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Re-verify master password to get a step-up token for sensitive operations",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Step-up authentication",
                "parameters": [
                    {
                        "description": "Auth hash for re-authentication",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.StepUpRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.StepUpResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.ErrorBody"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.ErrorBody"
                        }
                    }
                }
            }
        },
        "/auth/totp/disable": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deactivate two-factor authentication. Requires step-up.",
                "tags": [
                    "Auth"
                ],
                "summary": "Disable TOTP 2FA",
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "403": {
                        "description": "Step-up required",
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.ErrorBody"
                        }
                    }
                }
            }
        },
        "/auth/totp/enable": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Verify a TOTP code to activate two-factor authentication",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Enable TOTP 2FA",
                "parameters": [
                    {
                        "description": "TOTP verification code",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.TOTPCodeRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "400": {
                        "description": "Invalid code",
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.ErrorBody"
                        }
                    }
                }
            }
        },
        "/auth/totp/setup": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Generate a new TOTP secret and QR code URL. Requires step-up authentication.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Setup TOTP 2FA",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.TOTPSetupResponse"
                        }
                    },
                    "403": {
                        "description": "Step-up required",
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.ErrorBody"
                        }
                    }
                }
            }
        },
        "/auth/totp/verify": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Validate a TOTP code during the two-factor authentication step of login",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Verify TOTP code",
                "parameters": [
                    {
                        "description": "TOTP code",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.TOTPCodeRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "400": {
                        "description": "Invalid code",
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.ErrorBody"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_presenters_api.ErrorBody": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "string"
                        },
                        "field": {
                            "type": "string"
                        },
                        "message": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "internal_presenters_api.ListBackupsResponse": {
            "type": "object",
            "properties": {
                "backups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_presenters_api.BackupInfoDTO"
                    }
                }
            }
        },
        "internal_presenters_api.LoginRequest": {
            "type": "object",
            "properties": {
                "authHash": {
                    "description": "redacted via LOG_REDACT_FIELDS (C4)",
                    "type": "string"
                },
                "deviceName": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                }
            }
        },
...
        "internal_presenters_api.LogoutRequest": {
            "type": "object",
            "properties": {
                "refreshToken": {
                    "type": "string"
                }
            }
        },
...
        "internal_presenters_api.RedeemInviteRequest": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "internal_presenters_api.RedeemInviteResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "orgId": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter \"Bearer {token}\" (include the word Bearer)",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}